Move main to launcher
and change dependencies to runtimeOnly
