+++ prototypes/javaspec-launcher/build.gradle
@@ -11,8 +11,10 @@ application {
 }
 
 dependencies {
-//  testImplementation 'org.junit.jupiter:junit-jupiter-api:5.8.1'
-//  testRuntimeOnly 'org.junit.jupiter:junit-jupiter-engine:5.8.1'
+  implementation 'org.junit.platform:junit-platform-launcher:1.8.1'
+
+  runtimeOnly project(':javaspec-client')
+  runtimeOnly project(':javaspec-engine')
 }
 
 repositories {
